requirements-cpu.txt
# Зависимости для сервера БЕЗ GPU (CPU only)
# Использование: pip install -r requirements-cpu.txt

torch==2.9.1
torchaudio==2.9.1
torchvision==0.24.1
accelerate==1.12.0
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
faiss-cpu==1.13.0
filelock==3.20.0
fonttools==4.61.0
frozenlist==1.8.0
fsspec==2025.10.0
ftfy==6.3.1
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
Jinja2==3.1.6
kiwisolver==1.4.9
magic-filter==1.0.12
MarkupSafe==3.0.3
matplotlib==3.10.7
mpmath==1.3.0
multidict==6.7.0
networkx==3.6
numpy==2.2.6
open_clip_torch==3.2.0
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
packaging==25.0
pillow==12.0.0
polars==1.35.2
polars-runtime-32==1.35.2
propcache==0.4.1
psutil==7.1.3
pydantic==2.7.4
pydantic_core==2.18.4
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-telegram-bot==22.5
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
safetensors==0.7.0
scipy==1.16.3
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sympy==1.14.0
timm==1.0.22
tokenizers==0.22.1
transformers==4.57.3
typer-slim==0.20.0
typing_extensions==4.15.0
ultralytics==8.3.233
ultralytics-thop==2.0.18
urllib3==2.5.0
wcwidth==0.2.14
yarl==1.22.0
